test: authentication test

+++ database/migrations/2018_08_14_055151_create_i_classes_table.php
@@ -33,24 +33,26 @@ class CreateIClassesTable extends Migration
 
         });
 
-        // create the history table
-        Schema::dropIfExists('i_class_history');
-        DB::unprepared("CREATE TABLE i_class_history LIKE i_classes;");
-        // alter table
-        DB::unprepared("ALTER TABLE i_class_history MODIFY COLUMN id int(11) NOT NULL, 
+        if(DB::getDriverName() === 'mysql') {
+            // create the history table
+            Schema::dropIfExists('i_class_history');
+            DB::unprepared("CREATE TABLE i_class_history LIKE i_classes;");
+            // alter table
+            DB::unprepared("ALTER TABLE i_class_history MODIFY COLUMN id int(11) NOT NULL, 
    DROP PRIMARY KEY, ENGINE = MyISAM, ADD action VARCHAR(8) DEFAULT 'insert' FIRST, 
    ADD revision INT(6) NOT NULL AUTO_INCREMENT AFTER action,
    ADD PRIMARY KEY (id, revision);");
 
-        DB::unprepared("DROP TRIGGER IF EXISTS i_class__ai;");
-        DB::unprepared("DROP TRIGGER IF EXISTS i_class__au;");
-        //create after insert trigger
-        DB::unprepared("CREATE TRIGGER i_class__ai AFTER INSERT ON i_classes FOR EACH ROW
+            DB::unprepared("DROP TRIGGER IF EXISTS i_class__ai;");
+            DB::unprepared("DROP TRIGGER IF EXISTS i_class__au;");
+            //create after insert trigger
+            DB::unprepared("CREATE TRIGGER i_class__ai AFTER INSERT ON i_classes FOR EACH ROW
     INSERT INTO i_class_history SELECT 'insert', NULL, d.* 
     FROM i_classes AS d WHERE d.id = NEW.id;");
-        DB::unprepared("CREATE TRIGGER i_class__au AFTER UPDATE ON i_classes FOR EACH ROW
+            DB::unprepared("CREATE TRIGGER i_class__au AFTER UPDATE ON i_classes FOR EACH ROW
     INSERT INTO i_class_history SELECT 'update', NULL, d.*
     FROM i_classes AS d WHERE d.id = NEW.id;");
+        }
 
 
     }
@@ -62,9 +64,11 @@ class CreateIClassesTable extends Migration
      */
     public function down()
     {
-        DB::unprepared("DROP TRIGGER IF EXISTS i_class__ai;");
-        DB::unprepared("DROP TRIGGER IF EXISTS i_class__au;");
-        Schema::dropIfExists('i_class_history');
+        if(DB::getDriverName() === 'mysql') {
+            DB::unprepared("DROP TRIGGER IF EXISTS i_class__ai;");
+            DB::unprepared("DROP TRIGGER IF EXISTS i_class__au;");
+            Schema::dropIfExists('i_class_history');
+        }
         Schema::dropIfExists('i_classes');
     }
 }
