+++ database/migrations/2019_01_12_151224_create_subjects_table.php
@@ -29,36 +29,42 @@ class CreateSubjectsTable extends Migration
             $table->foreign('class_id')->references('id')->on('i_classes');
         });
 
-        // create the history table
-        Schema::dropIfExists('subject_history');
-        DB::unprepared("CREATE TABLE subject_history LIKE subjects;");
-        // alter table
-        DB::unprepared("ALTER TABLE subject_history MODIFY COLUMN id int(11) NOT NULL, 
-   DROP PRIMARY KEY, ENGINE = MyISAM, ADD action VARCHAR(8) DEFAULT 'insert' FIRST, 
-   ADD revision INT(6) NOT NULL AUTO_INCREMENT AFTER action,
-   ADD PRIMARY KEY (id, revision);");
+        if(DB::getDriverName() === 'mysql') {
+            // create the history table
+            Schema::dropIfExists('subject_history');
+            DB::unprepared("CREATE TABLE subject_history LIKE subjects;");
+            // alter table
+            DB::unprepared("ALTER TABLE subject_history MODIFY COLUMN id int(11) NOT NULL, 
+        DROP PRIMARY KEY, ENGINE = MyISAM, ADD action VARCHAR(8) DEFAULT 'insert' FIRST, 
+        ADD revision INT(6) NOT NULL AUTO_INCREMENT AFTER action,
+        ADD PRIMARY KEY (id, revision);");
 
-        DB::unprepared("DROP TRIGGER IF EXISTS subject_ai;");
-        DB::unprepared("DROP TRIGGER IF EXISTS subject_au;");
-        //create after insert trigger
-        DB::unprepared("CREATE TRIGGER subject_ai AFTER INSERT ON subjects FOR EACH ROW
-    INSERT INTO subject_history SELECT 'insert', NULL, d.* 
-    FROM subjects AS d WHERE d.id = NEW.id;");
-        DB::unprepared("CREATE TRIGGER subject_au AFTER UPDATE ON subjects FOR EACH ROW
-    INSERT INTO subject_history SELECT 'update', NULL, d.*
-    FROM subjects AS d WHERE d.id = NEW.id;");
+            DB::unprepared("DROP TRIGGER IF EXISTS subject_ai;");
+            DB::unprepared("DROP TRIGGER IF EXISTS subject_au;");
+            //create after insert trigger
+            DB::unprepared("CREATE TRIGGER subject_ai AFTER INSERT ON subjects FOR EACH ROW
+        INSERT INTO subject_history SELECT 'insert', NULL, d.* 
+        FROM subjects AS d WHERE d.id = NEW.id;");
+            DB::unprepared("CREATE TRIGGER subject_au AFTER UPDATE ON subjects FOR EACH ROW
+        INSERT INTO subject_history SELECT 'update', NULL, d.*
+        FROM subjects AS d WHERE d.id = NEW.id;");
+        }
+
+
     }
 
     /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        DB::unprepared("DROP TRIGGER IF EXISTS subject_ai;");
-        DB::unprepared("DROP TRIGGER IF EXISTS subject_au;");
-        Schema::dropIfExists('subject_history');
+        * Reverse the migrations.
+        *
+        * @return void
+        */
+        public function down()
+        {
+        if(DB::getDriverName() === 'mysql') {
+            DB::unprepared("DROP TRIGGER IF EXISTS subject_ai;");
+            DB::unprepared("DROP TRIGGER IF EXISTS subject_au;");
+            Schema::dropIfExists('subject_history');
+        }
         Schema::dropIfExists('subjects');
-    }
+        }
 }
